Split out gm_16AE
lbl_8046B6A0 needs to be the start of a file's bss section, so this
fixes some matches and allows me to continue matching these.
