+++ config/GALE01/splits.txt
@@ -1670,13 +1670,21 @@ melee/gm/gmmain.c:
 	.sbss       start:0x804D6590 end:0x804D6598
 
 melee/gm/gm_1601.c:
-	.text       start:0x801601C4 end:0x80174274
-	.rodata     start:0x803B75F8 end:0x803B7B18
-	.data       start:0x803D4C38 end:0x803D6488
-	.bss        start:0x8046B378 end:0x8046DBE8
+	.text       start:0x801601C4 end:0x8016AE38
+	.rodata     start:0x803B75F8 end:0x803B7A60
+	.data       start:0x803D4C38 end:0x803D5620
+	.bss        start:0x8046B378 end:0x8046B6A0
 	.sdata      start:0x804D3EF8 end:0x804D3F68
-	.sbss       start:0x804D6598 end:0x804D65B8
-	.sdata2     start:0x804DA230 end:0x804DA300
+	.sbss       start:0x804D6598 end:0x804D65A0
+	.sdata2     start:0x804DA230 end:0x804DA298
+
+melee/gm/gm_16AE.c:
+	.text       start:0x8016AE38 end:0x80174274
+	.rodata     start:0x803B7A60 end:0x803B7B18
+	.data       start:0x803D5620 end:0x803D6488
+	.bss        start:0x8046B6A0 end:0x8046DBE8
+	.sbss       start:0x804D65A0 end:0x804D65B8
+	.sdata2     start:0x804DA298 end:0x804DA300
 
 melee/gm/gmresult.c:
 	.text       start:0x80174274 end:0x80177724
