+++ src/melee/gm/gmresult.c
@@ -1,6 +1,6 @@
 #include "gm_1601.h"
 
-#include "gm_1601.static.h"
+#include "gm_16AE.static.h"
 
 
 int fn_80174274(void)
