+++ src/melee/gm/gmresultplayer.c
@@ -1,6 +1,6 @@
 #include "gm_1601.h"
 
-#include "gm_1601.static.h"
+#include "gm_16AE.static.h"
 
 
 /// #gm_80177724
